requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.3
crispy-bootstrap4==2022.1
cryptography==41.0.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-bootstrap5==23.3
django-crispy-forms==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==8.0.0
executing==1.2.0
filelock==3.12.2
frozenlist==1.4.0
idna==3.4
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
magic-filter==1.0.11
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2023.7.23
platformdirs==3.5.3
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.7
psycopg2-binary==2.9.7
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.1
virtualenv-clone==0.5.7
wcwidth==0.2.6
yarl==1.9.2
